+++ lib/generators/bootsy/install_generator.rb
@@ -1,5 +1,7 @@
 module Bootsy
   module Generators
+    # Public: A Rails Generator to copy translations, assets and routes to
+    # the host app.
     class InstallGenerator < Rails::Generators::Base
       source_root File.expand_path('../templates', __FILE__)
 
@@ -8,42 +10,43 @@ module Bootsy
       end
 
       def copy_locale
-        copy_file '../../../../config/locales/bootsy.en.yml', 'config/locales/bootsy.en.yml'
+        copy_file '../../../../config/locales/bootsy.en.yml',
+                  'config/locales/bootsy.en.yml'
       end
 
-      def add_assets
-        [
-          {
-            original: 'app/assets/javascripts/application.js',
-            skip_if: 'require bootsy',
-            content: "\n//= require bootsy",
-            position: {
-              after: '//= require jquery_ujs'
-            }
-          },
-          {
-            original: 'app/assets/stylesheets/application.css',
-            skip_if: 'require bootsy',
-            content: "\n *= require bootsy",
-            position: {
-              after: '*= require_self'
-            }
-          }
-        ].each do |params|
-          if File.exists?(params[:original])
-            if File.binread(params[:original]).include?(params[:skip_if])
-              say_status('skipped', "insert into #{params[:original]}", :yellow)
-            else
-              insert_into_file(params[:original], params[:content], params[:position])
-            end
-          else
-            say_status('not found', "#{params[:original]} not found. You must manually require Bootsy in your assets pipeline.", :red)
-          end
-        end
+      def add_javascript
+        require_asset(
+          'app/assets/javascripts/application.js',
+          "\n//= require bootsy",
+          '//= require jquery_ujs'
+        )
+      end
+
+      def add_stylesheet
+        require_asset(
+          'app/assets/stylesheets/application.css',
+          "\n *= require bootsy",
+          '*= require_self'
+        )
       end
 
       def copy_config
         template 'bootsy.rb', 'config/initializers/bootsy.rb'
+      end
+
+      private
+
+      def require_asset(destination, content, after_line)
+        if File.exist?(destination)
+          if File.binread(destination).include?(content)
+            say_status('skipped', "insert into #{destination}", :yellow)
+          else
+            insert_into_file(destination, content, after: after_line)
+          end
+        else
+          say_status('not found', "#{destination} not found. You must
+            manually require Bootsy in your assets pipeline.", :red)
+        end
       end
     end
   end